Isolate compatibility morph regressions behind bounded test-only guards
