+++ docs/architecture.md
@@ -826,6 +826,8 @@ Expensive geometry/text/image/mesh realization should be represented as immutabl
 
 Morph lowering may prepare immutable endpoint geometry in a fixed rendering coordinate frame, with a derived render transform separate from the effective semantic transform. This specialization must preserve world-space point correspondence, stroke behavior, channel precedence, and spatial queries. The current screen-space specialization requires finite, safe endpoint geometry and an invertible interpolated scale; unsupported cases retain current-relative evaluation. Independent affine drivers release the fixed frame before taking ownership, and exact morph endpoints restore source/target content without a stale render override. Native execution shares the compiled geometry directly; a genuine worker boundary installs its session-scoped resources once and sends references with per-frame progress, rather than reconstructing or retransmitting paths on each tick.
 
+Until #959 retires the existing animated compatibility input, its morph transport regressions may exercise that input only in the explicitly test-only `retained_execution_resources/morph_tests.rs` and `retained_resource_transport/morph_tests.rs` child modules. `scripts/retained-dynamic-stress-perf.mjs` is an explicit export/performance fixture for the unchanged gallery source, not a new authoring or execution authority. The architecture ratchet records exact shrinking token budgets for these fixtures, verifies their test-only module ownership, and continues to reject new production consumers. Remove these allowances with the compatibility input.
+
 Replacing persistent semantic content prepares a new `ResourceVersion`, validates it against the relevant semantic/source revision, then publishes the new reference only as part of a coherent semantic/execution publication. Failed or stale preparation leaves the previous live version intact.
 
 Runtime/effective content drivers, where supported for compatibility such as redraw-style behavior, must use an explicit effective resource/publication path rather than implicitly rewriting authored semantic content every frame. They remain subject to resource-version, stale-result and retirement rules, but do not create a `SceneRevision` merely because the effective resource changed.
